nightly update.

+++ solutions/kubernetes/ceph-deploy-functions.sh
@@ -412,41 +412,45 @@ function io_operation() {
 }
 
 function destroy_cluster_vm() {
-add_secondary_separator "Stop all ceph daemon"
-systemctl stop ceph.target
+if ! which ceph; then
+    add_secondary_separator "Ceph is not installed"
+else
+    add_secondary_separator "Stop all ceph daemon"
+    systemctl stop ceph.target
 
-add_secondary_separator "Remove cluster"
-fsid=$(cat /etc/ceph/ceph.conf | grep fsid | awk '{ print $3 }')
-echo "$fsid"
-cephadm rm-cluster --fsid $fsid --force
+    add_secondary_separator "Remove cluster"
+    fsid=$(cat /etc/ceph/ceph.conf | grep fsid | awk '{ print $3 }')
+    echo "$fsid"
+    cephadm rm-cluster --fsid $fsid --force
 
-add_secondary_separator "Zap OSDs"
-cephadm zap-osds --fsid $fsid --force
+    add_secondary_separator "Zap OSDs"
+    cephadm zap-osds --fsid $fsid --force
 
-add_secondary_separator "Uninstall Ceph Packages"
-dnf repository-packages Ceph remove -y
+    add_secondary_separator "Uninstall Ceph Packages"
+    dnf repository-packages Ceph remove -y
 
-add_secondary_separator "Unmount osd tmpfs"
-osd_mount=$(df -hT | grep osd | awk '{ print $7}')
-echo "OSD mounts to unmount: $osd_mount"
-for mount in osd_mount;	do
-	umount mount
-done
+    add_secondary_separator "Unmount osd tmpfs"
+    osd_mount=$(df -hT | grep osd | awk '{ print $7}')
+    echo "OSD mounts to unmount: $osd_mount"
+    for mount in osd_mount;	do
+        umount mount
+    done
 
-add_secondary_separator "Remove files"
-files_to_remove=(
-    "/tmp/monmap"
-    "/etc/yum.repos.d/ceph.repo"
-    "/var/lib/ceph"
-    "/lib/systemd/system/ceph*"
-    "/etc/yum.repos.d/_copr\:copr.fedorainfracloud.org\:tchaikov\:python*"
-)
-for file in ${files_to_remove[@]}; do
-    if [ -f "$file" ] || [ -d "$file" ]; then
-        echo "Removing file/folder $file"
-        rm -rf $file
-    fi
-done
+    add_secondary_separator "Remove files"
+    files_to_remove=(
+        "/tmp/monmap"
+        "/etc/yum.repos.d/ceph.repo"
+        "/var/lib/ceph"
+        "/lib/systemd/system/ceph*"
+        "/etc/yum.repos.d/_copr\:copr.fedorainfracloud.org\:tchaikov\:python*"
+    )
+    for file in ${files_to_remove[@]}; do
+        if [ -f "$file" ] || [ -d "$file" ]; then
+            echo "Removing file/folder $file"
+            rm -rf $file
+        fi
+    done
+fi
 }
 
 function destroy_cluster_docker() {
